+++ packages/bytes-to-co2/rollup.config.js
@@ -1,0 +1,3 @@
+import config from '../../rollup.config';
+
+export default config({ input: 'src/index.ts', name: 'co2Data' });