+++ packages/bytes-to-co2/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.8 (2020-10-24)
+
+**Note:** Version bump only for package bytes-to-co2
+
+
+
+
+
+## 0.3.7 (2020-10-24)
+
+**Note:** Version bump only for package bytes-to-co2
+
+
+
+
+
+## 0.3.6 (2020-10-24)
+
+**Note:** Version bump only for package bytes-to-co2
+
+
+
+
+
+## 0.3.5 (2020-10-24)
+
+**Note:** Version bump only for package bytes-to-co2
+
+
+
+
+
 ## 0.3.4 (2020-10-24)
 
 **Note:** Version bump only for package bytes-to-co2